+++ samples/advanced-samples/extensions-example/src/main/java/com/zhuinden/simplestackextensionsample/utils/RxRelayUtils.kt
@@ -1,0 +1,11 @@
+package com.zhuinden.simplestackextensionsample.utils
+
+import com.jakewharton.rxrelay2.BehaviorRelay
+
+fun <T> BehaviorRelay<T>.get(): T = value!!
+
+fun <T> BehaviorRelay<T>.getOrNull(): T? = value
+
+fun <T> BehaviorRelay<T>.set(value: T) {
+    this.accept(value)
+}